+++ app.js
@@ -1,47 +1,7 @@
 const http = require('http');
-const fs = require('fs');
+const routes = require('./routes');
 
-const server = http.createServer((req, res) => {
-    
-    const url = req.url;
-    const method = req.method;
-
-    if (url === '/') {
-        res.write('<html>');
-        res.write('<head><title>First website!</title></head>');
-        res.write('<body><form action="/message" method="POST"><input type="text" name="message"><button type="submit">Send</button></form></body>');
-        res.write('</html>');
-        return res.end();
-    }
-
-    if (url === '/message' && method === 'POST') {
-        const body = [];
-
-        req.on('data', (chunk) => {
-            console.log(chunk);
-            body.push(chunk);
-        });
-
-        req.on('end', () => {
-            const parsedBody = Buffer.concat(body).toString();
-            const message = parsedBody.split('=')[1];
-            console.log(parsedBody);
-            console.log(message);
-            fs.writeFile('message.txt', message, err => {
-                res.statusCode = 302;
-                res.setHeader('Location', '/');
-                return res.end();
-            });
-        });
-    }
-
-    res.setHeader('Content-Type', 'text/html');
-    res.write('<html>');
-    res.write('<head><title>First website!</title></head>');
-    res.write('<body><h1>Welcome to my first website in node.js!</h1></body>');
-    res.write('</html>');
-    res.end();
-});
+const server = http.createServer(routes.handler);
 
 // visit localhost:3000 in browser, to see my first website! :)
 server.listen(3000);
