+++ message.txt
@@ -1,1 +1,1 @@
-TEST
+TEASF
