+++ routes.js
@@ -1,0 +1,54 @@
+
+const fs = require('fs');
+
+const requestHandler = (req, res) => {
+
+    const url = req.url;
+    const method = req.method;
+
+    if (url === '/') {
+        res.write('<html>');
+        res.write('<head><title>First website!</title></head>');
+        res.write('<body><form action="/message" method="POST"><input type="text" name="message"><button type="submit">Send</button></form></body>');
+        res.write('</html>');
+        return res.end();
+    }
+
+    if (url === '/message' && method === 'POST') {
+        const body = [];
+
+        req.on('data', (chunk) => {
+            console.log(chunk);
+            body.push(chunk);
+        });
+
+        return req.on('end', () => {
+            const parsedBody = Buffer.concat(body).toString();
+            const message = parsedBody.split('=')[1];
+            console.log(parsedBody);
+            console.log(message);
+            fs.writeFile('message.txt', message, err => {
+                res.statusCode = 302;
+                res.setHeader('Location', '/');
+                return res.end();
+            });
+        });
+    }
+
+    res.setHeader('Content-Type', 'text/html');
+    res.write('<html>');
+    res.write('<head><title>First website!</title></head>');
+    res.write('<body><h1>Welcome to my first website in node.js!</h1></body>');
+    res.write('</html>');
+    res.end();
+};
+
+// module.export = requestHandler;
+
+// module.exports.handles = requestHandler;
+
+// exports.handler = requestHandler;
+
+module.exports = {
+    handler: requestHandler,
+};
